TemplateCache/customs: don't fail on missing keys
If 'customs' returns an error for missing keys on changes,
it's difficult to add a new custom key with unreleased changes pending
and a template that uses 'customs'.

Change 'customs' to yield an empty string for a missing custom key.
This matches the behavior of accessing the missing key directly
via `$change.Custom.Foo`.

Resolves #536