Fix slashes in FragmentFileFormat
Fixes https://github.com/miniscruff/changie/issues/559.

If the template for FragmentFileFormat resulted in any slashes being
present in the final string the system would treat that as a nested
directory under the "unreleased" directory.

This fixes that so that if slashes are in the final string they are
first replaced with dashes to make a safe filename.